🗑️ Remove buggy issue template
This template was initially supposed to be for users who use bedrockpy to be mentioned on the documentation. However GitHub was infinitely loading so i had to rewrite it multiple times until i was finally able to commit. However it seems like just the boilerplate template has been saved and nothing else.